requirements.txt
pandas==2.3.0
numpy==2.2.6
psycopg2-binary==2.9.10
PyMySQL==1.1.1
SQLAlchemy==2.0.41
python-dotenv==1.1.0
python-json-logger==3.3.0
